moved entity files into packages
