+++ app-server/src/main/java/com/inventory/app/server/controller/CollectionController.java
@@ -1,8 +1,8 @@
 package com.inventory.app.server.controller;
 
 import com.google.common.base.Preconditions;
-import com.inventory.app.server.entity.CollectionDetails;
-import com.inventory.app.server.entity.Media;
+import com.inventory.app.server.entity.collection.CollectionDetails;
+import com.inventory.app.server.entity.media.Media;
 import com.inventory.app.server.error.ResourceNotFoundException;
 import com.inventory.app.server.service.collection.CollectionDetailsService;
 import com.inventory.app.server.utility.RestPreConditions;
