+++ app-server/src/main/java/com/inventory/app/server/entity/collection/BaseCollection.java
@@ -1,4 +1,4 @@
-package com.inventory.app.server.entity;
+package com.inventory.app.server.entity.collection;
 
 
 import jakarta.persistence.*;
