+++ app-server/src/main/java/com/inventory/app/server/entity/media/Game.java
@@ -1,4 +1,4 @@
-package com.inventory.app.server.entity;
+package com.inventory.app.server.entity.media;
 
 import jakarta.persistence.*;
 import lombok.Getter;
