+++ app-server/src/main/java/com/inventory/app/server/entity/media/Movie.java
@@ -1,4 +1,5 @@
-package com.inventory.app.server.entity;
+package com.inventory.app.server.entity.media;
+
 
 import lombok.Getter;
 import lombok.Setter;
@@ -6,20 +7,17 @@ import lombok.Setter;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
 import jakarta.persistence.ElementCollection;
-import jakarta.persistence.Column;
 import jakarta.persistence.FetchType;
-
+import jakarta.persistence.Column;
 import java.time.LocalDate;
 import java.util.List;
-@Entity(name = "TelevisionShow")
-@Table(name = "televisiion_show")
+@Entity(name = "Media")
+@Table(name = "movie")
 @Getter
 @Setter
-public class TelevisionShow extends Media {
+public class Movie extends Media {
     @ElementCollection(fetch = FetchType.LAZY)
-    private List<String> writers;
-    @Column(name = "season")
-    private Integer season;
+    private List<String> directors;
     @Column(name = "release_date")
     private LocalDate releaseDate;
 }
