+++ app-server/src/main/java/com/inventory/app/server/entity/media/TelevisionShow.java
@@ -1,4 +1,4 @@
-package com.inventory.app.server.entity;
+package com.inventory.app.server.entity.media;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -11,15 +11,15 @@ import jakarta.persistence.FetchType;
 
 import java.time.LocalDate;
 import java.util.List;
-@Entity(name = "Music")
-@Table(name = "music")
+@Entity(name = "TelevisionShow")
+@Table(name = "televisiion_show")
 @Getter
 @Setter
-public class Music extends Media {
+public class TelevisionShow extends Media {
     @ElementCollection(fetch = FetchType.LAZY)
-    private List<String> artists;
-    @ElementCollection(fetch = FetchType.LAZY)
-    private List<String> songList;
+    private List<String> writers;
+    @Column(name = "season")
+    private Integer season;
     @Column(name = "release_date")
     private LocalDate releaseDate;
 }
