+++ app-server/src/main/java/com/inventory/app/server/mapper/BookMapper.java
@@ -1,7 +1,7 @@
 package com.inventory.app.server.mapper;
 
 import com.inventory.app.server.config.MediaInventoryAdditionalAttributes;
-import com.inventory.app.server.entity.Book;
+import com.inventory.app.server.entity.media.Book;
 import com.inventory.app.server.entity.payload.request.MediaId;
 import com.inventory.app.server.entity.payload.request.MediaRequest;
 import com.inventory.app.server.entity.payload.response.MediaResponse;
